Move event handling to 'util' subpackage
