+++ src/pyama/session/model.py
@@ -10,9 +10,9 @@ import pandas as pd
 import skimage.morphology as skmorph
 
 from . import const
-from .events import Event
-from .status import DummyStatus
 
+from ..util.events import Event
+from ..util.status import DummyStatus
 from ..io import StackdataIO
 from ..roi import ContourRoi
 from ..stack import Stack
